Converted multisig rotation command line coroutine to be unlinked to command line functionality but instead runs by processing inputs on a buffer.
